modularize helper functions for testing
